+++ generator/CHANGELOG.md
@@ -19,11 +19,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ---
 
+## [0.3.0] - 2025-05-16
+
+### Added
+
+- Added code to generate complex data types (work-in-progress)
+
+### Changed
+
+- Updated dependencies
+- Minor updates to CodeSystem Enum class generator
+- Minor updates to TypeScript data model generator
+- Changed approach for generated barrel files from a single top-level file to directory-level files
+
+
 ## [0.2.0] - 2025-05-01
 
 ### Added
 
-- Created the initial code to generate data models and write them to disk in the approprriate NPM package.
+- Created the initial code to generate data models and write them to disk in the appropriate NPM package.
   Only the selected CodeSystem Enum classes are currently generated.
 - Added unit tests for all current code.