+++ packages/fhir-core/CHANGELOG.md
@@ -19,6 +19,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ---
 
+## [0.3.0] - 2025-05-16
+
+### Changed
+
+- Updated dependencies
+- Changed export of utility/json-helpers to a named export
+
 ## [0.2.0] - 2025-05-01
 
 ### Changed